Added MemeService - used by AddMemeModule to save and by custom message handler to load commands.
