+++ SteakBot.Core/EventHandlers/CustomMessageHandlers/CommandMessageHandlers/CustomCommandMessageHandler.cs
@@ -1,10 +1,8 @@
 ﻿using System.Collections.Generic;
-using System.Configuration;
-using System.IO;
 using System.Linq;
 using Discord;
 using Discord.WebSocket;
-using Newtonsoft.Json;
+using SteakBot.Core.Modules;
 using SteakBot.Core.Objects;
 using SteakBot.Core.Objects.Enums;
 
@@ -12,48 +10,17 @@ namespace SteakBot.Core.EventHandlers.CustomMessageHandlers.CommandMessageHandle
 {
     public  class CustomCommandMessageHandler : BaseCommandMessageHandler
     {
-        private static readonly string MemeCommandsFileName = ConfigurationManager.AppSettings["memeCommandsRelativeFilePath"];
-        private static readonly string MemeCommandsOriginRelativeFilePath = ConfigurationManager.AppSettings["memeCommandsOriginRelativeFilePath"];
+        private readonly MemeService _memeService;
 
-        internal static IList<MemeCommand> Commands { get; set; } = LoadCommands();
+        internal static IList<MemeCommand> Commands { get; set; }
 
-        public CustomCommandMessageHandler()
+        public CustomCommandMessageHandler(MemeService memeService)
         {
+            _memeService = memeService;
+            _memeService.ReloadCommandsEvent += ReloadCommands;
+
+            Commands = _memeService.LoadCommands();
             CommandNames = Commands.Select(x => x.Name);
-        }
-
-        public bool SaveCommand(MemeCommand newCmd)
-        {
-            if (ValidateNewCommand(newCmd))
-            {
-                Commands.Add(newCmd);
-                using (var fileWriter = new StreamWriter(MemeCommandsFileName))
-                {
-                    using (var jsonTextWriter = new JsonTextWriter(fileWriter))
-                    {
-                        var jsonSerializer = new JsonSerializer();
-                        jsonSerializer.Serialize(jsonTextWriter, Commands);
-                    }
-                }
-
-#if DEBUG
-                // Replace original file.
-                using (var fileWriter = new StreamWriter(MemeCommandsOriginRelativeFilePath))
-                {
-                    using (var jsonTextWriter = new JsonTextWriter(fileWriter))
-                    {
-                        var jsonSerializer = new JsonSerializer();
-                        jsonSerializer.Serialize(jsonTextWriter, Commands);
-                    }
-                }
-#endif
-
-                ReloadCommands();
-
-                return true;
-            }
-
-            return false;
         }
 
         protected override bool InvokeInner(SocketUserMessage message)
@@ -89,32 +56,10 @@ namespace SteakBot.Core.EventHandlers.CustomMessageHandlers.CommandMessageHandle
 
         #region Private methods
 
-        private static IList<MemeCommand> LoadCommands()
-        {
-            using (var fileReader = new StreamReader(MemeCommandsFileName))
-            {
-                using (var jsonTextReader = new JsonTextReader(fileReader))
-                {
-                    var jsonSerializer = new JsonSerializer();
-                    return jsonSerializer.Deserialize<IList<MemeCommand>>(jsonTextReader);
-                }
-            }
-        }
-
         private void ReloadCommands()
         {
-            Commands = LoadCommands();
+            Commands = _memeService.LoadCommands();
             CommandNames = Commands.Select(x => x.Name);
-        }
-
-        private static bool ValidateNewCommand(MemeCommand newCmd)
-        {
-            if (Commands.Contains(newCmd))
-            {
-                return false;
-            }
-
-            return true;
         }
 
         #endregion
